RDKINCDT-21871: Fix nss package dependency
Reason for change: It seems that the networkmanager is being enabled for the first time in release 1.17.0. This module has a dependency on nss, which in turn depends on nss-native to support its ipk postinstall script. However, since the image assembler does not have a direct dependency on networkmanager and its ipks are being installed via the middleware packagegroup (packagegroup-middleware-layer), the nss-native is not being built and the necessary utility (shlibsign) is not made available for the do_rootfs task.